This PR has the following changes:
- Slot class is removed. In all use cases except `lower_graph` we really
just needed the attribute name and thus having an extra layer of
abstraction through Slot only made the code harder to understand.
- get_parameters, get_attributes, get_modules, and get_slots now return
a list of <name, item> pairs instead of a list of Slots.

Differential Revision: [D17728910](https://our.internmc.facebook.com/intern/diff/D17728910)

[ghstack-poisoned]

+++ caffe2/python/layers/sparse_lookup.py
@@ -550,6 +550,6 @@ class SparseLookup(ModelLayer):
         if not self.support_8bit() and version in {'uint8rowwise',
                                                    'fused_uint8rowwise',
                                                    'fused_uint4rowwise'}:
-            version = 'fp16'
+            version = 'fp32'
 
         self._add_ops(net, version)
